ContactFix

+++ ContactUs/models.py
@@ -3,9 +3,8 @@ from django.db import models
 class ContactData(models.Model):
     name = models.CharField(max_length=100)
     email = models.EmailField()
-    contact = models.IntegerField()
+    contact = models.BigIntegerField()
     message = models.TextField()
 
     def __str__(self):
         return self.name
-
